ios fix pt3
