--- src/main/java/com/solvd/demoapp/components/quantity/Quantity.java
@@ -1,10 +1,0 @@
-package com.solvd.demoapp.components.quantity;
-
-import org.openqa.selenium.SearchContext;
-import org.openqa.selenium.WebDriver;
-
-public class Quantity extends QuantityBase {
-    public Quantity(WebDriver driver, SearchContext searchContext) {
-        super(driver, searchContext);
-    }
-}
