+++ src/main/java/com/solvd/demoapp/components/quantity/QuantityBase.java
@@ -2,7 +2,6 @@ package com.solvd.demoapp.components.quantity;
 
 import com.zebrunner.carina.webdriver.decorator.ExtendedWebElement;
 import com.zebrunner.carina.webdriver.gui.AbstractUIObject;
-import com.zebrunner.carina.webdriver.locator.ExtendedFindBy;
 import lombok.Getter;
 import org.openqa.selenium.SearchContext;
 import org.openqa.selenium.WebDriver;
@@ -13,43 +12,13 @@ import org.slf4j.LoggerFactory;
 public abstract class QuantityBase extends AbstractUIObject {
     private static final Logger LOGGER = LoggerFactory.getLogger(QuantityBase.class);
 
-    @ExtendedFindBy(accessibilityId = "AddPlus Icons")
-    private ExtendedWebElement plusButton;
-
-    @ExtendedFindBy(accessibilityId = "SubtractMinus Icons")
-    private ExtendedWebElement minusButton;
-
-    @ExtendedFindBy(accessibilityId = "Amount")
-    private ExtendedWebElement amount;
-
     public QuantityBase(WebDriver driver, SearchContext searchContext) {
         super(driver, searchContext);
     }
 
-    public boolean clickPlus(int number) {
-        LOGGER.info("clickPlus(" + number + ")");
-        String value = amount.getAttribute("value");
-        return buttonClick(number, value, plusButton) - Integer.parseInt(value) == number;
-    }
+    public abstract boolean clickPlus(int number);
 
-    public boolean clickMinus(int number) {
-        LOGGER.info("clickMinus(" + number + ")");
-        String value = amount.getAttribute("value");
-        if (number > Integer.parseInt(value)) {
-            throw new IllegalArgumentException("Number must be less than existing quantity");
-        }
-        return buttonClick(number, value, minusButton) + number == Integer.parseInt(value);
-    }
+    public abstract boolean clickMinus(int number);
 
-    private Integer buttonClick(int number, String value, ExtendedWebElement extendedWebElement) {
-        LOGGER.info("buttonClick()");
-        if (number < 0) {
-            throw new IllegalArgumentException("Number must be grater that 0");
-        }
-        for (int i = 0; i < number; i++) {
-            extendedWebElement.click();
-        }
-        String changedValue = amount.getAttribute("value");
-        return Integer.parseInt(changedValue);
-    }
+    public abstract ExtendedWebElement getAmount();
 }
